# api/swagger/swagger.yaml
openapi: 3.0.0
info:
  version: "2.0.0"
  title: CCXT REST
  contact:
    name: "Adroit"
    email: "hello@adroit.ph"
    url: "https://adroit.ph/ccxt-rest-contact-us/"
  license:
    name: "MIT"
    url: "https://github.com/franz-see/ccxt-rest/blob/master/LICENSE.txt"
servers:
  - url: 'http://localhost:3000/'
...
paths:
...
  /exchange/{exchangeName}:
    get:
      description: Get the exchange details given the exchangeName and access token in the header
      tags:
        - Exchange Management API
      x-swagger-router-controller: exchange
      operationId: getExchange
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeResponse'
        "404":
          description: Exchange with that name is NOT supported or Exchange with that id does NOT exist
        "500":
          description: If an unexpected error occurred
    post:
      description: Creates an instance of the exchange
      tags:
        - Exchange Management API
      operationId: createExchange
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
      requestBody:
        description: The exchange to create.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/exchangeConfig'
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessToken'
        "404":
          description: Exchange with that name is NOT supported
        "500":
          description: If an unexpected error occurred
        "503":
          description: Support for exchange is currently broken
      x-swagger-router-controller: exchange
    delete:
      description: Delete the exchange details referenced by access token in the header
      security:
        - bearerAuth: []
      tags:
        - Exchange Management API
      x-swagger-router-controller: exchange
      operationId: deleteExchange
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeResponse'
        "404":
          description: Exchange with that name is NOT supported
        "500":
          description: If an unexpected error occurred
  '/exchange/{exchangeName}/{exchangeId}/markets':
    get:
      description: >-
        Get the markets of the exchange referenced by the exchangeName and exchangeId. 
        
        
        <br/> *Parameters listed here are common to all exchanges. But any other parameter passed would be forwarded as well into the exchange.*
      tags:
        - Public Data API
      operationId: markets
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
        - $ref: "#/components/parameters/exchangeIdParam"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MarketResponse'
                x-content-type: application/json
        "400":
          description: If the exchange itself complained about the parameters passed
        "401":
          description: If the exchange integration requires api key and secret for this function
        "403":
          description: If the exchange integration had an authentication issue (most probably nonce error)
        "404":
          description: Exchange with that name is NOT supported or Exchange with that id does NOT exist
        "500":
          description: If an unexpected error occurred
        "501":
          description: If the exchange integration does NOT support this function
        "504":
          description: >-
            If the exchange itself could not be reached because of some network
            error
      x-swagger-router-controller: exchange
  '/exchange/{exchangeName}/{exchangeId}/orderBook':
    get:
      description: >-
        Get the order book of the exchange referenced by the exchangeName, exchangeId and symbol. 
        
        
        <br/> *Parameters listed here are common to all exchanges. But any other parameter passed would be forwarded as well into the exchange.*
      tags:
        - Public Data API
      x-swagger-router-controller: exchange
      operationId: orderBook
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
        - $ref: "#/components/parameters/exchangeIdParam"
        - $ref: "#/components/parameters/requiredExchangeSymbolParam"
        - name: limit
          in: query
          description: The limit of the exchange's order book to be retrieved.
          required: false
          style: form
          explode: true
          schema:
            type: number
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderBookResponse'
                x-content-type: application/json
        "400":
          description: If the exchange itself complained about the parameters passed
        "401":
          description: If the exchange integration requires api key and secret for this function
        "403":
          description: If the exchange integration had an authentication issue (most probably nonce error)
        "404":
          description: Exchange with that name is NOT supported or Exchange with that id does NOT exist
        "500":
          description: If an unexpected error occurred
        "501":
          description: If the exchange integration does NOT support this function
        "504":
          description: If the exchange itself could not be reached because of some network error
  '/exchange/{exchangeName}/{exchangeId}/l2OrderBook':
    get:
      description: >-
        Get the Level 2 Order Book of the exchange referenced by the exchangeName, exchangeId and symbol. 
        
        
        <br/> *Parameters listed here are common to all exchanges. But any other parameter passed would be forwarded as well into the exchange.*
      tags:
        - Public Data API
      x-swagger-router-controller: exchange
      operationId: l2OrderBook
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
        - $ref: "#/components/parameters/exchangeIdParam"
        - $ref: "#/components/parameters/requiredExchangeSymbolParam"
        - name: limit
          in: query
          description: The limit of the exchange's order book to be retrieved.
          required: false
          style: form
          explode: true
          schema:
            type: number
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderBookResponse'
                x-content-type: application/json
        "400":
          description: If the exchange itself complained about the parameters passed
        "401":
          description: If the exchange integration requires api key and secret for this function
        "403":
          description: If the exchange integration had an authentication issue (most probably nonce error)
        "404":
          description: Exchange with that name is NOT supported or Exchange with that id does NOT exist
        "500":
          description: If an unexpected error occurred
        "501":
          description: If the exchange integration does NOT support this function
        "504":
          description: If the exchange itself could not be reached because of some network error
...
  '/exchange/{exchangeName}/{exchangeId}/tickers':
    get:
      description: >-
        Get the tickers of the exchange referenced by the exchangeName, exchangeId and symbol. 
        
        
        <br/> *Parameters listed here are common to all exchanges. But any other parameter passed would be forwarded as well into the exchange.*
      tags:
        - Public Data API
      x-swagger-router-controller: exchange
      operationId: tickers
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
        - $ref: "#/components/parameters/exchangeIdParam"
        - $ref: "#/components/parameters/optionalExchangeSymbolParam"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TickerResponse'
                x-content-type: application/json
        "400":
          description: If the exchange itself complained about the parameters passed
        "401":
          description: If the exchange integration requires api key and secret for this function
        "403":
          description: If the exchange integration had an authentication issue (most probably nonce error)
        "404":
          description: Exchange with that name is NOT supported or Exchange with that id does NOT exist
        "500":
          description: If an unexpected error occurred
        "501":
          description: If the exchange integration does NOT support this function
        "504":
          description: If the exchange itself could not be reached because of some network error
  /exchange/{exchangeName}/{exchangeId}/balances:
    get:
      description: >-
        Get the balances of the exchange referenced by the exchangeName and exchangeId. 
        
        
        <br/> *Parameters listed here are common to all exchanges. But any other parameter passed would be forwarded as well into the exchange.*
      tags:
        - Private Data API
      x-swagger-router-controller: exchange
      operationId: balances
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
        - $ref: "#/components/parameters/exchangeIdParam"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        "400":
          description: If the exchange itself complained about the parameters passed
        "401":
          description: If the exchange integration requires api key and secret for this function
        "403":
          description: If the exchange integration had an authentication issue (most probably nonce error)
        "404":
          description: Exchange with that name is NOT supported or Exchange with that id does NOT exist
        "500":
          description: If an unexpected error occurred
        "501":
          description: If the exchange integration does NOT support this function
        "504":
          description: If the exchange itself could not be reached because of some network error
  '/exchange/{exchangeName}/{exchangeId}/orders':
    get:
      description: >-
        Get the orders of the exchange referenced by the exchangeName and exchangeId. 
        
        
        <br/> *Parameters listed here are common to all exchanges. But any other parameter passed would be forwarded as well into the exchange.*
      tags:
        - Private Data API
      x-swagger-router-controller: exchange
      operationId: fetchOrders
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
        - $ref: "#/components/parameters/exchangeIdParam"
        - $ref: "#/components/parameters/optionalExchangeSymbolParam"
        - name: since
          in: query
          description: Retrieve the orders starting from 'since'
          required: false
          style: form
          explode: true
          schema:
            type: string
        - name: limit
          in: query
          description: The limit of the exchange's orders to be retrieved.
          required: false
          style: form
          explode: true
          schema:
            type: number
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponse'
                x-content-type: application/json
        "400":
          description: If the exchange itself complained about the parameters passed
        "401":
          description: If the exchange integration requires api key and secret for this function
        "403":
          description: If the exchange integration had an authentication issue (most probably nonce error)
        "404":
          description: Exchange with that name is NOT supported or Exchange with that id does NOT exist
        "500":
          description: If an unexpected error occurred
        "501":
          description: If the exchange integration does NOT support this function
        "504":
          description: If the exchange itself could not be reached because of some network error
  /exchange/{exchangeName}/{exchangeId}/orders/open:
    get:
      description: >-
        Get the open orders of the exchange referenced by the exchangeName and exchangeId. 
        
        
        <br/> *Parameters listed here are common to all exchanges. But any other parameter passed would be forwarded as well into the exchange.*
      tags:
        - Private Data API
      x-swagger-router-controller: exchange
      operationId: fetchOpenOrders
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
        - $ref: "#/components/parameters/exchangeIdParam"
        - $ref: "#/components/parameters/optionalExchangeSymbolParam"
        - name: since
          in: query
          description: Retrieve the orders starting from 'since'
          required: false
          style: form
          explode: true
          schema:
            type: string
        - name: limit
          in: query
          description: The limit of the exchange's orders to be retrieved.
          required: false
          style: form
          explode: true
          schema:
            type: number
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponse'
                x-content-type: application/json
        "400":
          description: If the exchange itself complained about the parameters passed
        "401":
          description: If the exchange integration requires api key and secret for this function
        "403":
          description: If the exchange integration had an authentication issue (most probably nonce error)
        "404":
          description: Exchange with that name is NOT supported or Exchange with that id does NOT exist
        "500":
          description: If an unexpected error occurred
        "501":
          description: If the exchange integration does NOT support this function
        "504":
          description: If the exchange itself could not be reached because of some network error
  /exchange/{exchangeName}/{exchangeId}/orders/closed:
    get:
      description: >-
        Get the closed orders of the exchange referenced by the exchangeName and exchangeId. 
        
        
        <br/> *Parameters listed here are common to all exchanges. But any other parameter passed would be forwarded as well into the exchange.*
      tags:
        - Private Data API
      x-swagger-router-controller: exchange
      operationId: fetchClosedOrders
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
        - $ref: "#/components/parameters/exchangeIdParam"
        - $ref: "#/components/parameters/optionalExchangeSymbolParam"
        - name: since
          in: query
          description: Retrieve the orders starting from 'since'
          required: false
          style: form
          explode: true
          schema:
            type: string
        - name: limit
          in: query
          description: The limit of the exchange's orders to be retrieved.
          required: false
          style: form
          explode: true
          schema:
            type: number
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponse'
                x-content-type: application/json
        "400":
          description: If the exchange itself complained about the parameters passed
        "401":
          description: If the exchange integration requires api key and secret for this function
        "403":
          description: If the exchange integration had an authentication issue (most probably nonce error)
        "404":
          description: Exchange with that name is NOT supported or Exchange with that id does NOT exist
        "500":
          description: If an unexpected error occurred
        "501":
          description: If the exchange integration does NOT support this function
        "504":
          description: If the exchange itself could not be reached because of some network error
  /exchange/{exchangeName}/{exchangeId}/trades/mine:
    get:
      description: >-
        Get my trades of the exchange referenced by the exchangeName and exchangeId. 
        
        
        <br/> *Parameters listed here are common to all exchanges. But any other parameter passed would be forwarded as well into the exchange.*
      tags:
        - Private Data API
      x-swagger-router-controller: exchange
      operationId: fetchMyTrades
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
        - $ref: "#/components/parameters/exchangeIdParam"
        - $ref: "#/components/parameters/optionalExchangeSymbolParam"
        - name: since
          in: query
          description: Retrieve the trades starting from 'since'
          required: false
          style: form
          explode: true
          schema:
            type: string
        - name: limit
          in: query
          description: The limit of the exchange's trades to be retrieved.
          required: false
          style: form
          explode: true
          schema:
            type: number
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TradeResponse'
                x-content-type: application/json
        "400":
          description: If the exchange itself complained about the parameters passed
        "401":
          description: If the exchange integration requires api key and secret for this function
        "403":
          description: If the exchange integration had an authentication issue (most probably nonce error)
        "404":
          description: Exchange with that name is NOT supported or Exchange with that id does NOT exist
        "500":
          description: If an unexpected error occurred
        "501":
          description: If the exchange integration does NOT support this function
        "504":
          description: If the exchange itself could not be reached because of some network error
  /exchange/{exchangeName}/{exchangeId}/order:
    post:
      description: Create an order on the exchange referenced by the exchangeName and exchangeId
      tags:
        - Private Data API
      x-swagger-router-controller: exchange
      operationId: createOrder
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
        - $ref: "#/components/parameters/exchangeIdParam"
      requestBody:
        description: The order to place.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/orderPlacement'
        required: false
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        "400":
          description: If the exchange itself complained about the parameters passed
        "401":
          description: If the exchange integration requires api key and secret for this function
        "403":
          description: If the exchange integration had an authentication issue (most probably nonce error)
        "404":
          description: Exchange with that name is NOT supported or Exchange with that id does NOT exist
        "500":
          description: If an unexpected error occurred
        "501":
          description: If the exchange integration does NOT support this function
        "504":
          description: If the exchange itself could not be reached because of some network error
  /exchange/{exchangeName}/{exchangeId}/order/{orderId}:
    get:
      description: >-
        Retrieves the information of an order on the exchange referenced by the exchangeName, exchangeId and orderId. 
        
        
        <br/> *Parameters listed here are common to all exchanges. But any other parameter passed would be forwarded as well into the exchange.*
      tags:
        - Private Data API
      x-swagger-router-controller: exchange
      operationId: fetchOrder
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
        - $ref: "#/components/parameters/exchangeIdParam"
        - $ref: "#/components/parameters/orderIdParam"
        - $ref: "#/components/parameters/optionalExchangeSymbolParam"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        "400":
          description: If the exchange itself complained about the parameters passed
        "401":
          description: If the exchange integration requires api key and secret for this function
        "403":
          description: If the exchange integration had an authentication issue (most probably nonce error)
        "404":
          description: Exchange with that name is NOT supported or Exchange with that id does NOT exist
        "500":
          description: If an unexpected error occurred
        "501":
          description: If the exchange integration does NOT support this function
        "504":
          description: If the exchange itself could not be reached because of some network error
    delete:
      description: >- 
        Cancel an open order on the exchange referenced by the exchangeName, exchangeId and orderId. 
        
        
        <br/> *Parameters listed here are common to all exchanges. But any other parameter passed would be forwarded as well into the exchange.*
      tags:
        - Private Data API
      x-swagger-router-controller: exchange
      operationId: cancelOrder
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
        - $ref: "#/components/parameters/exchangeIdParam"
        - $ref: "#/components/parameters/orderIdParam"
        - $ref: "#/components/parameters/optionalExchangeSymbolParam"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        "400":
          description: If the exchange itself complained about the parameters passed
        "401":
          description: If the exchange integration requires api key and secret for this function
        "403":
          description: If the exchange integration had an authentication issue (most probably nonce error)
        "404":
          description: Exchange with that name is NOT supported or Exchange with that id does NOT exist
        "500":
          description: If an unexpected error occurred
        "501":
          description: If the exchange integration does NOT support this function
        "504":
          description: If the exchange itself could not be reached because of some network error
...
components:
  schemas:
...
    AccessToken:
      required:
        - token
      properties:
        token:
          type: string
          description: >- 
            The JWT token that needs to be added into the 'Authorization'
            header with the 'Bearer ' prefix. For example, given a token of xyz,
            do a secured request with 'Authorization: Bearer xyz'
    ExchangeResponse:
      required:
        - id
        - countries
        - enableRateLimit
        - name
        - private
      type: object
      properties:
        id:
          type: string
          description: >-
            The id of the exchange. When you created the exchange (via
            POST:/exchanges/{exhangeName}), the 'id' parameter there becomes the
            name here
        name:
          type: string
          description: >-
            The name of the exchange. 
        private:
          type: boolean
          description: Whether this exchange is private (has apiKey) or public (no apiKey)
        enableRateLimit:
          type: boolean
          description: >-
            Whether to enable the built in rate limiter or not. The built in
            rate limiter is an approximation of the actual exchange's limit. To
            have a more accurate rate limiting, set this to false and implement
            the rate limiter on your client
          default: true
        countries:
          type: array
          description: The list of countries where this exchange is a member of
          items:
            type: string
        rateLimit:
          type: integer
          description: >-
            A request rate limit in milliseconds. Specifies the required minimal
            delay between two consequent HTTP requests to the same exchange. If
            enableRateLimit is set to false, this would be ignored.
        twofa:
          type: boolean
          description: Whether to enable two factor authentication or not
          default: false
        has:
          $ref: '#/components/schemas/ExchangeHasCapabilities'
        urls:
          type: object
          properties: {}
          description: Collection of URLs this exchange has
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT  
  parameters:
...
    exchangeNameParam:
      name: exchangeName
      in: path
      description: >-
        The name of the exchange. Possible values are any of the result of
        GET:/exchanges.
      required: true
      style: simple
      explode: false
      schema:
        $ref : "#/components/schemas/Exchange"